web: refactor 1/7 data processing
